Session participant check in SessionCtrl
Also moved participant check functions into SessionCtrl
