+++ router/api/sockets.js
@@ -5,7 +5,6 @@ const socket = require('socket.io')
 
 const config = require('../../config.js')
 const SessionCtrl = require('../../controllers/SessionCtrl.js')
-const helpers = require('./helpers.js')
 
 // Create an HTTPS server if in production, otherwise use HTTP.
 const createServer = app => {
@@ -102,7 +101,7 @@ module.exports = function (app) {
         },
         function (err, session) {
           // Don't let anyone but the session's student or volunteer create messages
-          if (err || helpers.isNotSessionParticipant(session, data.user)) {
+          if (err || SessionCtrl.isNotSessionParticipant(session, data.user)) {
             console.log('Could not deliver message')
             io.emit('error', err || 'Only session participants are allowed to send messages')
             return
@@ -127,7 +126,7 @@ module.exports = function (app) {
 
     socket.on('drawClick', function (data) {
       if (!data || !data.sessionId) return
-      helpers.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
+      SessionCtrl.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
         if (err) return
         socket.broadcast.to(data.sessionId).emit('draw', {
           x: data.x,
@@ -139,7 +138,7 @@ module.exports = function (app) {
 
     socket.on('saveImage', function (data) {
       if (!data || !data.sessionId) return
-      helpers.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
+      SessionCtrl.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
         if (err) return
         socket.broadcast.to(data.sessionId).emit('save')
       })
@@ -147,7 +146,7 @@ module.exports = function (app) {
 
     socket.on('undoClick', function (data) {
       if (!data || !data.sessionId) return
-      helpers.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
+      SessionCtrl.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
         if (err) return
         socket.broadcast.to(data.sessionId).emit('undo')
       })
@@ -155,7 +154,7 @@ module.exports = function (app) {
 
     socket.on('clearClick', function (data) {
       if (!data || !data.sessionId) return
-      helpers.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
+      SessionCtrl.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
         if (err) return
         io.to(data.sessionId).emit('clear')
       })
@@ -163,7 +162,7 @@ module.exports = function (app) {
 
     socket.on('drawing', function (data) {
       if (!data || !data.sessionId) return
-      helpers.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
+      SessionCtrl.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
         if (err) return
         socket.broadcast.to(data.sessionId).emit('draw')
       })
@@ -171,7 +170,7 @@ module.exports = function (app) {
 
     socket.on('end', function (data) {
       if (!data || !data.sessionId) return
-      helpers.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err, session) {
+      SessionCtrl.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err, session) {
         if (err) return
         session.saveWhiteboardUrl(data.whiteboardUrl)
         socket.broadcast.to(data.sessionId).emit('end', data)
@@ -180,7 +179,7 @@ module.exports = function (app) {
 
     socket.on('changeColor', function (data) {
       if (!data || !data.sessionId) return
-      helpers.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
+      SessionCtrl.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
         if (err) return
         socket.broadcast.to(data.sessionId).emit('color', data.color)
       })
@@ -188,7 +187,7 @@ module.exports = function (app) {
 
     socket.on('changeWidth', function (data) {
       if (!data || !data.sessionId) return
-      helpers.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
+      SessionCtrl.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
         if (err) return
         socket.broadcast.to(data.sessionId).emit('width', data.width)
       })
@@ -196,7 +195,7 @@ module.exports = function (app) {
 
     socket.on('dragStart', function (data) {
       if (!data || !data.sessionId) return
-      helpers.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
+      SessionCtrl.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
         if (err) return
         socket.broadcast.to(data.sessionId).emit('dstart', {
           x: data.x,
@@ -208,7 +207,7 @@ module.exports = function (app) {
 
     socket.on('dragAction', function (data) {
       if (!data || !data.sessionId) return
-      helpers.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
+      SessionCtrl.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
         if (err) return
         socket.broadcast.to(data.sessionId).emit('drag', {
           x: data.x,
@@ -220,7 +219,7 @@ module.exports = function (app) {
 
     socket.on('dragEnd', function (data) {
       if (!data || !data.sessionId) return
-      helpers.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
+      SessionCtrl.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
         if (err) return
         socket.broadcast.to(data.sessionId).emit('dend', {
           x: data.x,
@@ -232,7 +231,7 @@ module.exports = function (app) {
 
     socket.on('insertText', function (data) {
       if (!data || !data.sessionId) return
-      helpers.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
+      SessionCtrl.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
         if (err) return
         io.to(data.sessionId).emit('text', {
           text: data.text,
@@ -243,7 +242,7 @@ module.exports = function (app) {
     })
 
     socket.on('resetScreen', function (data) {
-      helpers.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
+      SessionCtrl.verifySessionParticipantBySessionId(data.sessionId, data.user, function (err) {
         if (err) return
         io.to(data.sessionId).emit('reset')
       })
